Making the backend get_task_request actually do stuff. added functionality to api.rs and added entities.

+++ server/src/database/task.rs
@@ -5,7 +5,7 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "task")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: u64,
     pub title: String,
     pub completed: bool,
     pub last_edited: chrono::NaiveDateTime,
